+++ src/Model/Table/PostsTable.php
@@ -24,6 +24,8 @@ namespace MeCms\Model\Table;
 
 use ArrayObject;
 use Cake\Cache\Cache;
+use Cake\Database\Schema\Table as Schema;
+use Cake\Datasource\EntityInterface;
 use Cake\Event\Event;
 use Cake\Network\Exception\InternalErrorException;
 use Cake\ORM\Entity;
@@ -32,6 +34,7 @@ use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use MeCms\Model\Entity\Post;
 use MeCms\Model\Table\AppTable;
+use MeCms\Model\Table\Traits\GetPreviewFromTextTrait;
 use MeCms\Model\Table\Traits\IsOwnedByTrait;
 use MeCms\Model\Table\Traits\NextToBePublishedTrait;
 
@@ -40,9 +43,19 @@ use MeCms\Model\Table\Traits\NextToBePublishedTrait;
  * @property \Cake\ORM\Association\BelongsTo $Categories
  * @property \Cake\ORM\Association\BelongsTo $Users
  * @property \Cake\ORM\Association\BelongsToMany $Tags
+ * @method \MeCms\Model\Entity\Post get($primaryKey, $options = [])
+ * @method \MeCms\Model\Entity\Post newEntity($data = null, array $options = [])
+ * @method \MeCms\Model\Entity\Post[] newEntities(array $data, array $options = [])
+ * @method \MeCms\Model\Entity\Post|bool save(\Cake\Datasource\EntityInterface $entity, $options = [])
+ * @method \MeCms\Model\Entity\Post patchEntity(\Cake\Datasource\EntityInterface $entity, array $data, array $options = [])
+ * @method \MeCms\Model\Entity\Post[] patchEntities($entities, array $data, array $options = [])
+ * @method \MeCms\Model\Entity\Post findOrCreate($search, callable $callback = null, $options = [])
+ * @mixin \Cake\ORM\Behavior\TimestampBehavior
+ * @mixin \Cake\ORM\Behavior\CounterCacheBehavior
  */
 class PostsTable extends AppTable
 {
+    use GetPreviewFromTextTrait;
     use IsOwnedByTrait;
     use LocatorAwareTrait;
     use NextToBePublishedTrait;
@@ -52,6 +65,20 @@ class PostsTable extends AppTable
      * @var string
      */
     public $cache = 'posts';
+
+    /**
+     * Alters the schema used by this table. This function is only called after
+     *  fetching the schema out of the database
+     * @param Cake\Database\Schema\TableSchema $schema TableSchema instance
+     * @return Cake\Database\Schema\TableSchema TableSchema instance
+     * @since 2.17.0
+     */
+    protected function _initializeSchema(Schema $schema)
+    {
+        $schema->columnType('preview', 'json');
+
+        return $schema;
+    }
 
     /**
      * Called after an entity has been deleted
@@ -115,6 +142,23 @@ class PostsTable extends AppTable
                 $data['tags'][$k]['tag'] = $tag;
             }
         }
+    }
+
+    /**
+     * Called before each entity is saved
+     * @param \Cake\Event\Event $event Event object
+     * @param \Cake\ORM\Entity $entity Entity object
+     * @param \ArrayObject $options Options
+     * @return void
+     * @since 2.17.0
+     * @uses MeCms\Model\Table\AppTable::beforeSave()
+     * @uses MeCms\Model\Table\Traits\GetPreviewFromTextTrait::getPreview()
+     */
+    public function beforeSave(Event $event, EntityInterface $entity, ArrayObject $options)
+    {
+        parent::beforeSave($event, $entity, $options);
+
+        $entity->preview = $this->getPreview($entity->text);
     }
 
     /**
@@ -199,14 +243,14 @@ class PostsTable extends AppTable
                 //  less chance to find a related post
                 foreach (array_reverse($tags) as $tag) {
                     $post = $this->find('active')
-                        ->select(['id', 'title', 'slug', 'text'])
+                        ->select(['id', 'title', 'slug', 'text', 'preview'])
                         ->matching('Tags', function ($q) use ($tag) {
                             return $q->where([sprintf('%s.id', $this->Tags->getAlias()) => $tag->id]);
                         })
                         ->where([sprintf('%s.id NOT IN', $this->getAlias()) => $exclude]);
 
                     if ($images) {
-                        $post->where([sprintf('%s.text LIKE', $this->getAlias()) => sprintf('%%%s%%', '<img')]);
+                        $post->where([sprintf('%s.preview IS NOT', $this->getAlias()) => null]);
                     }
 
                     $post = $post->first();