+++ tests/TestCase/Model/Entity/PostTest.php
@@ -96,6 +96,7 @@ class PostTest extends TestCase
     public function testNoAccessibleProperties()
     {
         $this->assertFalse($this->Post->isAccessible('id'));
+        $this->assertFalse($this->Post->isAccessible('preview'));
         $this->assertFalse($this->Post->isAccessible('modified'));
     }
 
@@ -105,7 +106,7 @@ class PostTest extends TestCase
      */
     public function testVirtualFields()
     {
-        $this->assertEquals(['preview', 'tags_as_string'], $this->Post->getVirtual());
+        $this->assertEquals(['tags_as_string'], $this->Post->getVirtual());
     }
 
     /**