+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # 2.x branch
+## 2.17 branch
+### 2.17.0
+* added `og:image:width` and `og:image:height` meta tags for previews;
+* added `preview` field to pages and posts tables, which contains url, width
+    and height of the preview, encoded with json;
+* added `preview` and `thumbnail` virtual fields to photos;
+* improved `KcFinderComponent`. Now the class has `_getDefaultConfig()` and
+    `initialize()` method;
+* improved `AuthComponent`. Now the class has the `initialize()` method;
+* `BaseUpdateShell` renamed as `BaseUpdateConsole`. This creates less confusion.
+
 ## 2.16 branch
 ### 2.16.1
 * managers can delete photos and photos albums;